refactoring + cleanup
