+++ components/index.ts
@@ -2,4 +2,5 @@ export * from "./Button";
 export * from "./colors";
 export { default as Header } from "./Header";
 export { default as NavBar } from "./NavBar";
+export { default as SpeakerCard } from "./SpeakerCard";
 export * from "./SpeakerCard";
