+++ pages/index.tsx
@@ -1,7 +1,6 @@
 import { FC } from "react";
 import styled from "styled-components";
-import NavBar from "../components/NavBar";
-import { navbarBlue } from "../components/colors";
+import { NavBar, navbarBlue } from "../components";
 
 const LandingPage: FC = () => (
   <View>
@@ -89,6 +88,7 @@ export const Logo = styled.img`
   flex-basis: 1280px;
   flex-shrink: 1;
   width: 1280px;
+  
   @media (orientation: portrait) {
     max-width: 100vw;
     margin-bottom: 10px;
